refactor: テストコードの配置場所変更
